requirements_dev.txt
coverage==3.7.1
mock==1.0.1
nose==1.3.3
nosegae==0.4.0
PyCrypto==2.6.0
webapp2==2.5.2
WebTest==2.0.18
